refactor: changed folder structure to follow best python conventions
