+++ Assignment1/lib/cmv.py
@@ -1,6 +1,6 @@
 import math
 import numpy as np
-from utils import *
+from lib.utils import *
 
 
 class CMV:
